Switch to javascript from typescript
