+++ server/package.json
@@ -1,9 +1,9 @@
 {
   "name": "ppi-canada-server",
   "version": "0.1.0",
-  "main": "src/server.ts",
+  "main": "src/server.js",
   "scripts": {
-    "dev": "ts-node-dev --respawn --transpile-only src/server.ts",
+    "dev": "ts-node-dev --respawn --transpile-only src/server.js",
     "build": "tsc",
     "start": "node dist/server.js"
   },
